Chomping pacman!

+++ Pacman/Actors/Ghost.hs
@@ -5,5 +5,5 @@ import Pacman.Actors.Base
 data Ghost = Ghost { position :: Vec2,
                      direction :: Direction }
 
-update :: Ghost -> Int -> Ghost
+update :: Ghost -> Float -> Ghost
 update ghost dt = ghost
